+++ alicloud/resource_alicloud_forward.go
@@ -1,11 +1,12 @@
 package alicloud
 
 import (
+	"time"
+
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/vpc"
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/terraform-providers/terraform-provider-alicloud/alicloud/connectivity"
-	"time"
 )
 
 func resourceAliyunForwardEntry() *schema.Resource {